// src/api/inventory/models/ServiceRequest.ts
export interface ServiceRequest {
    /**
     * 
     * @type {string}
     * @memberof ServiceRequest
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequest
     */
    description: string;
    /**
     * 
     * @type {Priority}
     * @memberof ServiceRequest
     */
    priority: Priority;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequest
     */
    equipmentId: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequest
     */
    id: string;
    /**
     * 
     * @type {ServiceRequestStatus}
     * @memberof ServiceRequest
     */
    status: ServiceRequestStatus;
    /**
     * 
     * @type {Equipment}
     * @memberof ServiceRequest
     */
    equipment?: Equipment;
    /**
     * 
     * @type {Date}
     * @memberof ServiceRequest
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ServiceRequest
     */
    updatedAt: Date;
}