// src/api/inventory/models/ServiceRequestCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Priority } from './Priority';
import {
    PriorityFromJSON,
    PriorityFromJSONTyped,
    PriorityToJSON,
} from './Priority';

/**
 * 
 * @export
 * @interface ServiceRequestCreate
 */
export interface ServiceRequestCreate {
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestCreate
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestCreate
     */
    description: string;
    /**
     * 
     * @type {Priority}
     * @memberof ServiceRequestCreate
     */
    priority: Priority;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestCreate
     */
    equipmentId: string;
}

/**
 * Check if a given object implements the ServiceRequestCreate interface.
 */
export function instanceOfServiceRequestCreate(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "title" in value;
    isInstance = isInstance && "description" in value;
    isInstance = isInstance && "priority" in value;
    isInstance = isInstance && "equipmentId" in value;

    return isInstance;
}

export function ServiceRequestCreateFromJSON(json: any): ServiceRequestCreate {
    return ServiceRequestCreateFromJSONTyped(json, false);
}

export function ServiceRequestCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceRequestCreate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': json['title'],
        'description': json['description'],
        'priority': PriorityFromJSON(json['priority']),
        'equipmentId': json['equipmentId'],
    };
}

export function ServiceRequestCreateToJSON(value?: ServiceRequestCreate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'description': value.description,
        'priority': PriorityToJSON(value.priority),
        'equipmentId': value.equipmentId,
    };
}