// src/api/inventory/models/ServiceRequestUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hospital Equipment & Service Management API
 * API for managing hospital equipment inventory, service requests, and location tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Priority } from './Priority';
import {
    PriorityFromJSON,
    PriorityFromJSONTyped,
    PriorityToJSON,
} from './Priority';
import type { ServiceRequestStatus } from './ServiceRequestStatus';
import {
    ServiceRequestStatusFromJSON,
    ServiceRequestStatusFromJSONTyped,
    ServiceRequestStatusToJSON,
} from './ServiceRequestStatus';

/**
 * 
 * @export
 * @interface ServiceRequestUpdate
 */
export interface ServiceRequestUpdate {
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestUpdate
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestUpdate
     */
    description?: string;
    /**
     * 
     * @type {Priority}
     * @memberof ServiceRequestUpdate
     */
    priority?: Priority;
    /**
     * 
     * @type {ServiceRequestStatus}
     * @memberof ServiceRequestUpdate
     */
    status?: ServiceRequestStatus;
}

/**
 * Check if a given object implements the ServiceRequestUpdate interface.
 */
export function instanceOfServiceRequestUpdate(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function ServiceRequestUpdateFromJSON(json: any): ServiceRequestUpdate {
    return ServiceRequestUpdateFromJSONTyped(json, false);
}

export function ServiceRequestUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ServiceRequestUpdate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': !exists(json, 'title') ? undefined : json['title'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'priority': !exists(json, 'priority') ? undefined : PriorityFromJSON(json['priority']),
        'status': !exists(json, 'status') ? undefined : ServiceRequestStatusFromJSON(json['status']),
    };
}

export function ServiceRequestUpdateToJSON(value?: ServiceRequestUpdate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'description': value.description,
        'priority': PriorityToJSON(value.priority),
        'status': ServiceRequestStatusToJSON(value.status),
    };
}